Move name generation to its own class
