+++ strawberry/schema/schema.py
@@ -117,7 +117,7 @@ class Schema(BaseSchema):
             (
                 directive
                 for directive in self.directives
-                if self.config.name_from_directive(directive) == graphql_name
+                if self.config.name_converter.from_directive(directive) == graphql_name
             ),
             None,
         )
